+++ src/components/WalletLayout/types.ts
@@ -4,6 +4,7 @@ type WalletMenuItem = {
   label: string;
   path: string;
   icon: ReactElement;
+  comingSoon?: boolean;
 };
 
 export type WalletLayoutProps = PropsWithChildren<{